+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - Function update: `sd_nav()` replacing `sd_next()`, which contains both previous and next buttons.
 - Function update: `sd_close()` now supports addition of previous button as well.
+- Function update: `sd_question()` now supports either `option` or `options`. If both are provided, `option` will be used. When defining `option` or `options`, users can now only provide display labels, and the db values will be auto-converted to snake case. For example, `sd_question(options = c("Option 1", "Option 2"))` will have db values of `option_1` and `option_2`.
 - New feature: `surveydown` now supports previous buttons. Under `survey-settings`, users can now define `show-previous` (default to `FALSE`) to enable previous button globally. In `sd_nav()`, users can define `show_previous` being `TRUE` or `FALSE` to override the global setting for that specific page.
 - Simplification: No need to define `format`, `echo`, and `warning` in YAML header of `survey.qmd` anymore. The survey can have a clean start and runs fine without the YAML header.
 - Simplification: No need to define `sd_nav()` for each page. It's auto-injected unless `sd_nav()` or `sd_close()` is explicitly defined in the page with or without parameters.